+++ CHANGELOG.md
@@ -10,6 +10,7 @@ If you like this project and find it useful, please consider giving it a star on
 
 - [shelly]: When an already discovered and stored device is discovered with a new address, the change is registered and takes effect after restarting.
 - [shelly]: Added support for shellypro3em.
+- [shelly]: Added a guide to add and debug BLU devices https://github.com/Luligu/matterbridge-shelly/blob/dev/BLU.md.
 
 ### Fixed